Math Quiz App
This was a final project I came up with for my Andriod Development app.
It is not 100% finished, but it does work and it is a good and easy way
to start. I wrote the code as simple as possible. Hopefully, one day
I'll have a finished version of the program. Enjoy.